+++ example/lib/increment/decrement-controller.dart
@@ -1,27 +1,16 @@
-import 'dart:async';
 
 import 'package:bloc_pattern/bloc_pattern.dart';
-import 'package:rxdart/rxdart.dart';
 
 class DecrementController extends BlocBase {
   DecrementController() {
     print("DecrementController create");
   }
 
-  //fluxo da variavel counter
-  var _counterController = BehaviorSubject<int>.seeded(0);
-  Stream<int> get outCounter => _counterController.stream;
-  Sink<int> get inCounter => _counterController.sink;
+  int counter = 0;
 
   void decrement() {
-    inCounter.add(_counterController.value - 1);
+    counter--;
+    notifyListeners();
   }
 
-  @override
-  void dispose() {
-    super.dispose();
-    print("DecrementController dispose");
-
-    _counterController.close();
-  }
 }